[CHORE] 사소한 코드 수정
