+++ src/main/java/com/hy/ouch/converter/UserConverter.java
@@ -3,7 +3,7 @@ package com.hy.ouch.converter;
 import org.springframework.stereotype.Component;
 
 import com.hy.ouch.domain.User;
-import com.hy.ouch.dto.user.response.MypageUserInfoResponse;
+import com.hy.ouch.dto.mypage.response.MypageUserInfoResponse;
 import com.hy.ouch.dto.user.response.UserInfoResponse;
 import com.hy.ouch.dto.user.response.UserSignupResponse;
