+++ src/main/java/com/hy/ouch/service/mypage/MypageService.java
@@ -1,0 +1,49 @@
+package com.hy.ouch.service.mypage;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.hy.ouch.converter.UserConverter;
+import com.hy.ouch.domain.Nation;
+import com.hy.ouch.domain.User;
+import com.hy.ouch.dto.mypage.request.MypageUserInfoUpdateRequest;
+import com.hy.ouch.dto.mypage.response.MypageUserInfoResponse;
+import com.hy.ouch.repository.nation.NationRepository;
+import com.hy.ouch.repository.user.UserRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class MypageService {
+
+	private final UserRepository userRepository;
+	private final NationRepository nationRepository;
+	private final UserConverter userConverter;
+
+	@Transactional(readOnly = true)
+	public MypageUserInfoResponse myPageGetUserInfo(Long userId) {
+		User user = userRepository.findById(userId)
+			.orElseThrow(() -> new RuntimeException("User not found"));
+		return userConverter.user2MypageUserInfoResponse(user);
+	}
+
+	@Transactional
+	public void myPageUpdateUserInfo(Long userId, MypageUserInfoUpdateRequest request) {
+		User user = userRepository.findById(userId)
+			.orElseThrow(() -> new RuntimeException("User not found"));
+
+		Nation wantedNation = nationRepository.findById(request.getNationId())
+			.orElseThrow(() -> new RuntimeException("Nation not found"));
+
+		User updatedUser = user.toBuilder()
+			.nickname(request.getNickname())
+			.phoneNumber(request.getPhoneNumber())
+			.gender(request.getGender())
+			.email(request.getEmail())
+			.nation(wantedNation)
+			.build();
+
+		userRepository.save(updatedUser);
+	}
+}
